% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edaphos_metrics.R
\name{edaphos_mae}
\alias{edaphos_mae}
\title{Mean absolute error}
\usage{
edaphos_mae(observed, predicted)
}
\arguments{
\item{observed, predicted}{Numeric vectors.}
}
\value{
Non-negative numeric scalar.
}
\description{
Mean absolute error
}
